+++ src/interpreter/InterpreterEngine.cpp
@@ -17,15 +17,15 @@
 #include "AggregateOp.h"
 #include "BinaryConstraintOps.h"
 #include "FunctorOps.h"
-#include "IOSystem.h"
+#include "io/IOSystem.h"
 #include "Logger.h"
 #include "ProfileEvent.h"
 #include "RamTypes.h"
-#include "ReadStream.h"
+#include "io/ReadStream.h"
 #include "RecordTable.h"
 #include "SignalHandler.h"
 #include "SymbolTable.h"
-#include "WriteStream.h"
+#include "io/WriteStream.h"
 #include "interpreter/InterpreterContext.h"
 #include "interpreter/InterpreterGenerator.h"
 #include "interpreter/InterpreterIndex.h"
