+++ src/io/WriteStream.h
