+++ src/io/WriteStreamSQLite.h
